requirements/requirements.txt
accelerate==0.18.0
clip-interrogator==0.6.0
diffusers==0.11.1
ftfy==6.1.1
Pillow==8.4.0
scipy==1.10.1
torch @ https://download.pytorch.org/whl/cu116/torch-1.13.1%2Bcu116-cp39-cp39-linux_x86_64.whl
torchaudio @ https://download.pytorch.org/whl/cu116/torchaudio-0.13.1%2Bcu116-cp39-cp39-linux_x86_64.whl
torchsummary==1.5.1
torchtext==0.14.1
torchvision @ https://download.pytorch.org/whl/cu116/torchvision-0.14.1%2Bcu116-cp39-cp39-linux_x86_64.whl
tqdm==4.65.0
transformers==4.27.3
